+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - fix: optimize increase/decrease number action.
 - feat: optimize search commands to support highlights.
 - feat: optimize move cursor actions to support visual mode.
+- feat: change case support visual mode selection.
 - chore: upgrade deps.
 
 ## v0.1.24